Separate mocks in different folders and update imports
