+++ test/mocks/player/currently-playing.mock.ts
