+++ test/mocks/search/search-artists.mock.ts
